+++ js/data.js
@@ -1,0 +1,8 @@
+import {createAdvert} from './utils.js';
+
+const ADVERTS_LENGTH = 10
+
+/* Массив из произвольных объектов недвижимости*/
+const adverts = new Array(ADVERTS_LENGTH).fill(null).map(() => createAdvert());
+
+export {adverts};
